% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEpanel.R
\name{DEpanel}
\alias{DEpanel}
\alias{DEpanelUI}
\alias{DEpanelServer}
\title{Generate the DE panel of the shiny app}
\usage{
DEpanelUI(id, metadata)

DEpanelServer(id, expression.matrix, metadata, anno)
}
\arguments{
\item{id}{the input slot that will be used to access the value}

\item{metadata}{a data frame containing metadata for the samples contained
in the expression.matrix; must contain at minimum two columns:
the first column must contain the column names of the expression.matrix,
while the last column is assumed to contain the experimental conditions
that will be tested for differential expression}

\item{expression.matrix}{the expression matrix; rows correspond to genes and
columns correspond to samples; usually preprocessed by
\code{\link{preprocessExpressionMatrix}}}

\item{anno}{annotation data frame containing a match between the row names
of the expression.matrix (usually ENSEMBL IDs) and the gene names that
should be rendered within the app and in output files; this object is
created by \code{\link{generateShinyApp}} using the org.db specified}
}
\description{
These are the UI and server components of the DE panel of the
shiny app. It is generated by including 'DE' in the panels.default argument
of \code{\link{generateShinyApp}}.
}
