% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplotPanel.R
\name{DEplotPanel}
\alias{DEplotPanel}
\alias{DEplotPanelUI}
\alias{DEplotPanelServer}
\title{Generate the DE plot plot panel of the shiny app}
\usage{
DEplotPanelUI(id)

DEplotPanelServer(id, DEresults, anno)
}
\arguments{
\item{id}{the input slot that will be used to access the value}

\item{DEresults}{differential expression results output from DEpanelServer;
a reactive list with slots 'DEtable' (all genes), 'DEtableSubset' (only DE genes),
'lfcThreshold' and 'pvalThreshold'}

\item{anno}{annotation data frame containing a match between the row names
of the expression.matrix (usually ENSEMBL IDs) and the gene names that
should be rendered within the app and in output files; this object is
created by \code{\link{generateShinyApp}} using the org.db specified}
}
\description{
These are the UI and server components of the DE plot panel of the
shiny app. It is generated by including 'DEplot' in the panels.default argument
of \code{\link{generateShinyApp}}.
}
