% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternFuns.R
\name{determine_uds}
\alias{determine_uds}
\title{Determine the pattern between two intervals}
\usage{
determine_uds(min1, max1, min2, max2)
}
\arguments{
\item{min1, max1, min2, max2}{the endpoints of the two intervals}
}
\value{
A single character (one of "U", "D", "S") representing the pattern
}
\description{
This function checks if the two input intervals oferlap and
outputs the corresponding pattern (up, down, or straight) based on that.
}
\examples{
determine_uds(10, 20, 15, 25) # overlap
determine_uds(10, 20, 25, 35) # no overlap
}
