% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCplotFuns.R
\name{jaccard_index}
\alias{jaccard_index}
\title{Calculate the Jaccard similarity index (JSI) between two vectors}
\usage{
jaccard_index(a, b)
}
\arguments{
\item{a, b}{two vectors}
}
\value{
The JSI of the two vectors, a single value between 0 and 1.
}
\description{
Calculate the Jaccard similarity index (JSI) between two vectors
}
\examples{
jaccard_index(1:4, 2:6)
}
