% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingPanel.R
\name{landingPanel}
\alias{landingPanel}
\alias{landingPanelUI}
\alias{landingPanelServer}
\title{Generate the landing page panel of the shiny app}
\usage{
landingPanelUI(id, show = TRUE)

landingPanelServer(id)
}
\arguments{
\item{id}{the input slot that will be used to access the value}

\item{show}{whether to show the panel or not; default is TRUE; there for
compatibility with specifying panels to show}
}
\value{
The UI and Server components of the shiny module, that can be used
within the UI and Server definitions of a shiny app.
}
\description{
These are the UI and server components of the landing page
panel of the shiny app. It is generated by including 'Landing' in the
panels.default argument of \code{\link{generateShinyApp}}.
}
