% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplotFuns.R
\name{ma_plot}
\alias{ma_plot}
\title{Create an MA plot visualising differential expression (DE) results}
\usage{
ma_plot(
  genes.de.results,
  pval.threshold = 0.05,
  lfc.threshold = 0.5,
  alpha = 0.1,
  ylims = NULL,
  add.colours = TRUE,
  add.expression.colour.gradient = TRUE,
  add.guide.lines = TRUE,
  add.labels.auto = TRUE,
  add.labels.custom = FALSE,
  ...
)
}
\arguments{
\item{genes.de.results}{the table of DE genes, usually generated by
\code{\link{DEanalysis_edger}}}

\item{pval.threshold}{the p-value and log2(fold-change)
thresholds to determine whether a gene is DE}

\item{lfc.threshold}{the p-value and log2(fold-change)
thresholds to determine whether a gene is DE}

\item{alpha}{the transparency of points; ignored for DE genes if
add.expression.colour.gradient is TRUE; default is 0.1}

\item{ylims}{a single value to create (symmetric) y-axis limits; by default
inferred from the data}

\item{add.colours}{whether to colour genes based on their log2(fold-change)
and -log10(p-value); default is TRUE}

\item{add.expression.colour.gradient}{whether to add a colour gradient
for DE genes to present their log2(expression); default is TRUE}

\item{add.guide.lines}{whether to add vertical and horizontal guide lines
to the plot to highlight the thresholds; default is TRUE}

\item{add.labels.auto}{whether to automatically label genes with the
highest |log2(fold-change)| and expression; default is TRUE}

\item{add.labels.custom}{whether to add labels to user-specified genes;
the parameter genes.to.label must also be specified; default is FALSE}

\item{...}{parameters passed on to \code{\link{ma_enhance}}}
}
\value{
The MA plot as a ggplot object.
}
\description{
This function creates an MA plot to visualise the results
of a DE analysis.
}
\seealso{
\code{\link{ma_enhance}}
}
