% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEsummaryFuns.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Create heatmap of an expression matrix}
\usage{
plot_heatmap(
  expression.matrix.subset,
  top.annotation.ids = NULL,
  metadata,
  type = "Log2 Expression"
)
}
\arguments{
\item{top.annotation.ids}{a vector of column indices denoting which columns
of the metadata should become heatmap annotations}

\item{metadata}{a data frame containing metadata for the samples contained
in the expression.matrix; must contain at minimum two columns:
the first column must contain the column names of the expression.matrix,
while the last column is assumed to contain the experimental conditions
that will be tested for differential expression}
}
\value{
The heatmap as detailed in the ComplexHeatmap package.
}
\description{
This function creates a heatmap to visualise an expression matrix
}
