% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelectPanel.R
\name{sampleSelectPanel}
\alias{sampleSelectPanel}
\alias{sampleSelectPanelUI}
\alias{sampleSelectPanelServer}
\title{Generate the sample select panel of the shiny app}
\usage{
sampleSelectPanelUI(id)

sampleSelectPanelServer(id, expression.matrix, metadata)
}
\arguments{
\item{id}{the input slot that will be used to access the value}

\item{expression.matrix}{the expression matrix; rows correspond to genes and
columns correspond to samples; usually preprocessed by
\code{\link{preprocessExpressionMatrix}}}

\item{metadata}{a data frame containing metadata for the samples contained
in the expression.matrix; must contain at minimum two columns:
the first column must contain the column names of the expression.matrix,
while the last column is assumed to contain the experimental conditions
that will be tested for differential expression}
}
\description{
These are the UI and server components of the sample selection
panel of the shiny app. It is generated by including 'SampleSelect' in the
panels.default argument of \code{\link{generateShinyApp}}.
}
