% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplotFuns.R
\name{volcano_enhance}
\alias{volcano_enhance}
\title{Add features to a volcano plot visualising differential expression (DE) results}
\usage{
volcano_enhance(
  vp,
  df,
  pval.threshold,
  lfc.threshold,
  alpha,
  add.colours,
  point.colours = c("#bfbfbf", "orange", "red", "blue"),
  raster = FALSE,
  add.expression.colour.gradient,
  colour.gradient.scale = list(left = c("#99e6ff", "#000066"), right = c("#99e6ff",
    "#000066")),
  colour.gradient.breaks = waiver(),
  colour.gradient.limits = NULL,
  add.guide.lines,
  guide.line.colours = c("green", "blue"),
  add.labels.auto,
  add.labels.custom,
  annotation = NULL,
  n.labels.auto = c(5, 5, 5),
  genes.to.label = NULL,
  seed = 0,
  label.force = 1
)
}
\arguments{
\item{vp}{volcano plot as a ggplot object (usually passed by \code{\link{volcano_plot}})}

\item{df}{data frame of DE results for all genes (usually passed by
\code{\link{volcano_plot}})}

\item{pval.threshold}{the p-value and log2(fold-change)
thresholds to determine whether a gene is DE}

\item{lfc.threshold}{the p-value and log2(fold-change)
thresholds to determine whether a gene is DE}

\item{alpha}{the transparency of points; ignored for DE genes if
add.expression.colour.gradient is TRUE; default is 0.1}

\item{add.colours}{whether to colour genes based on their log2(fold-change)
and -log10(p-value); default is TRUE}

\item{point.colours}{a vector of 4 colours to colour genes with both pval
and lfc under thresholds, just pval under threshold, just lfc under threshold,
both pval and lfc over threshold (DE genes) respectively; only used if
add.colours is TRUE}

\item{raster}{whether to rasterize non-DE genes with ggraster to reduce
memory usage; particularly useful when saving plots to files}

\item{add.expression.colour.gradient}{whether to add a colour gradient
for DE genes to present their log2(expression); default is TRUE}

\item{colour.gradient.scale}{a vector of two colours to create a colour
gradient for colouring the DE genes based on expression; a named list with
components left and right can be supplied to use two different colour scales;
only used if add.expression.colour.gradient is TRUE}

\item{colour.gradient.breaks, colour.gradient.limits}{parameters to customise
the legend of the colour gradient scale; especially useful if creating
multiple plots or a plot with two scales;
only used if add.expression.colour.gradient is TRUE}

\item{add.guide.lines}{whether to add vertical and horizontal guide lines
to the plot to highlight the thresholds; default is TRUE}

\item{guide.line.colours}{a vector with two colours to be used to colour
the guide lines; the first colour is used for the p-value and log2(fold-change)
thresholds and the second for double those values}

\item{add.labels.auto}{whether to automatically label genes with the
highest |log2(fold-change)| and expression; default is TRUE}

\item{add.labels.custom}{whether to add labels to user-specified genes;
the parameter genes.to.label must also be specified; default is FALSE}

\item{annotation}{annotation data frame containing a match between the gene
field of df (usually ENSEMBL IDs) and the gene names that should be shown
in the plot labels; not necessary if df already contains gene names}

\item{n.labels.auto}{a integer vector of length 3 denoting the number of
genes that should be automatically labelled; the first entry corresponds to
DE genes with the lowest p-value, the second to those with highest absolute
log2(fold-change) and the third to those with highest expression; a single
integer can also be specified, to be used for all 3 entries; default is 5}

\item{genes.to.label}{a vector of gene names (either matching the name field
of df or the corresponding name from the annotation) to be labelled in the
plot; if names are present those are shown as the labels (but the values are
the ones matched - this is to allow custom gene names to be presented)}

\item{seed}{the random seed to be used for reproducibility; only used for
ggrepel::geom_label_repel if labels are present}

\item{label.force}{passed to the force argument of ggrepel::geom_label_repel;
higher values make labels overlap less (at the cost of them being further
away from the points they are labelling)}
}
\value{
The enhanced volcano plot as a ggplot object.
}
\description{
This function adds features to a volcano plot and is usually
called indirectly by \code{\link{volcano_plot}}
}
\seealso{
\code{\link{volcano_plot}}
}
